Fix space lookup, extract space functionality into a separate file
