+++ pkg/storage/fs/owncloudsql/owncloudsql.go
@@ -54,7 +54,6 @@ import (
 	"github.com/cs3org/reva/pkg/storage/fs/registry"
 	"github.com/cs3org/reva/pkg/storage/utils/chunking"
 	"github.com/cs3org/reva/pkg/storage/utils/templates"
-	"github.com/cs3org/reva/pkg/utils"
 	"github.com/mitchellh/mapstructure"
 	"github.com/pkg/errors"
 	"github.com/pkg/xattr"
@@ -488,11 +487,6 @@ func (fs *owncloudsqlfs) getUserStorage(user string) (int, error) {
 		id, err = fs.filecache.CreateStorage("home::" + user)
 	}
 	return id, err
-}
-
-// CreateStorageSpace creates a storage space
-func (fs *owncloudsqlfs) CreateStorageSpace(ctx context.Context, req *provider.CreateStorageSpaceRequest) (*provider.CreateStorageSpaceResponse, error) {
-	return nil, errtypes.NotSupported("unimplemented: CreateStorageSpace")
 }
 
 func (fs *owncloudsqlfs) convertToResourceInfo(ctx context.Context, entry *filecache.File, ip string, mdKeys []string) (*provider.ResourceInfo, error) {
@@ -1952,113 +1946,6 @@ func (fs *owncloudsqlfs) HashFile(path string) (string, string, string, error) {
 
 		return string(sha1h.Sum(nil)), string(md5h.Sum(nil)), string(adler32h.Sum(nil)), nil
 	}
-}
-
-func (fs *owncloudsqlfs) ListStorageSpaces(ctx context.Context, filter []*provider.ListStorageSpacesRequest_Filter, _ map[string]struct{}) ([]*provider.StorageSpace, error) {
-	var (
-		spaceID = "*"
-		err     error
-	)
-
-	filteringUnsupportedSpaceTypes := false
-
-	for i := range filter {
-		switch filter[i].Type {
-		case provider.ListStorageSpacesRequest_Filter_TYPE_SPACE_TYPE:
-			t := filter[i].GetSpaceType()
-			filteringUnsupportedSpaceTypes = filteringUnsupportedSpaceTypes || (t != "personal" && !strings.HasPrefix(t, "+"))
-		case provider.ListStorageSpacesRequest_Filter_TYPE_ID:
-			spaceID, _, _ = utils.SplitStorageSpaceID(filter[i].GetId().OpaqueId)
-		}
-	}
-	if filteringUnsupportedSpaceTypes {
-		// owncloudsql only supports personal spaces, no need to search for something else
-		return []*provider.StorageSpace{}, nil
-	}
-
-	spaces := []*provider.StorageSpace{}
-	if spaceID == "*" {
-		spaces, err = fs.listAllPersonalSpaces()
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		space, err := fs.getSpaceByID(spaceID)
-		if err != nil {
-			return nil, err
-		}
-		spaces = append(spaces, space)
-	}
-	return spaces, nil
-}
-
-func (fs *owncloudsqlfs) listAllPersonalSpaces() ([]*provider.StorageSpace, error) {
-	storages, err := fs.filecache.ListStorages(true)
-	if err != nil {
-		return nil, err
-	}
-	spaces := []*provider.StorageSpace{}
-	for _, storage := range storages {
-		owner := strings.TrimPrefix(storage.ID, "home::")
-		u := &userpb.User{Id: &userpb.UserId{OpaqueId: owner}, Username: owner}
-		root, err := fs.filecache.Get(storage.NumericID, "")
-		if err != nil {
-			return nil, err
-		}
-		space := &provider.StorageSpace{
-			Id: &provider.StorageSpaceId{OpaqueId: strconv.Itoa(storage.NumericID)},
-			Root: &provider.ResourceId{
-				StorageId: strconv.Itoa(storage.NumericID),
-				OpaqueId:  strconv.Itoa(root.ID),
-			},
-			Name:      u.Username,
-			SpaceType: "personal",
-			Mtime:     &types.Timestamp{Seconds: uint64(root.MTime)},
-			Owner:     u,
-		}
-		spaces = append(spaces, space)
-	}
-	return spaces, nil
-}
-
-func (fs *owncloudsqlfs) getSpaceByID(spaceID string) (*provider.StorageSpace, error) {
-	storage, err := fs.filecache.GetStorage(spaceID)
-	if err != nil {
-		return nil, err
-	}
-	if !strings.HasPrefix(storage.ID, "home::") {
-		return nil, fmt.Errorf("only personal spaces are supported")
-	}
-
-	root, err := fs.filecache.Get(storage.NumericID, "")
-	if err != nil {
-		return nil, err
-	}
-	owner := strings.TrimPrefix(storage.ID, "home::")
-	u := &userpb.User{Id: &userpb.UserId{OpaqueId: owner}, Username: owner}
-
-	space := &provider.StorageSpace{
-		Id: &provider.StorageSpaceId{OpaqueId: strconv.Itoa(storage.NumericID)},
-		Root: &provider.ResourceId{
-			StorageId: strconv.Itoa(storage.NumericID),
-			OpaqueId:  strconv.Itoa(root.ID),
-		},
-		Name:      u.Username,
-		SpaceType: "personal",
-		Mtime:     &types.Timestamp{Seconds: uint64(root.MTime)},
-		Owner:     u,
-	}
-	return space, nil
-}
-
-// UpdateStorageSpace updates a storage space
-func (fs *owncloudsqlfs) UpdateStorageSpace(ctx context.Context, req *provider.UpdateStorageSpaceRequest) (*provider.UpdateStorageSpaceResponse, error) {
-	return nil, errtypes.NotSupported("update storage space")
-}
-
-// DeleteStorageSpace deletes a storage space
-func (fs *owncloudsqlfs) DeleteStorageSpace(ctx context.Context, req *provider.DeleteStorageSpaceRequest) error {
-	return errtypes.NotSupported("delete storage space")
 }
 
 func readChecksumIntoResourceChecksum(ctx context.Context, checksums, algo string, ri *provider.ResourceInfo) {
